requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
WTForms==3.1.2
email-validator==2.1.1